projects/python-url-shortener-microservice/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=22.0
drf-spectacular>=0.27
